discord-bots-python-backup-20251225/requirements.txt
# Discord Bot Dependencies

# Core Discord library
discord.py>=2.3.2

# Environment variables
python-dotenv>=1.0.0

# Async HTTP requests
aiohttp>=3.9.0

# YAML configuration
PyYAML>=6.0.1

# Database (optional - uncomment if needed)
# SQLAlchemy>=2.0.0
# aiosqlite>=0.19.0

# Web scraping (optional - uncomment if needed)
# beautifulsoup4>=4.12.0
# lxml>=4.9.0

# Music bots (optional - uncomment if needed)
# yt-dlp>=2023.10.0
# PyNaCl>=1.5.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
